PEP8: Renames Script modules
