+++ tmt_future/scripts/sync_photometry_video.py
